Using AbstractEntity works with postman now/refactored repositories into repository folders
