+++ TEAR-API/src/main/java/org/launchcode/TEAR_API/repositories/UserRepository.java
@@ -1,4 +1,4 @@
-package org.launchcode.TEAR_API.data;
+package org.launchcode.TEAR_API.repositories;
 
 import org.launchcode.TEAR_API.models.User;
 import org.springframework.data.repository.CrudRepository;
